Remove SPTPersistentCacheTypes
* Move the enums and foundation exports to the files
that own them
