+++ include/SPTPersistentCache/SPTPersistentCacheResponse.h
@@ -20,9 +20,29 @@
  */
 #import <Foundation/Foundation.h>
 
-#import <SPTPersistentCache/SPTPersistentCacheTypes.h>
-
 @class SPTPersistentCacheRecord;
+
+/**
+ * The SPTPersistentCacheResponseCode enum defines constants that is used to identify what kind of response would be
+ * given in callback to loadDataForKey:withCallback: method.
+ */
+typedef NS_ENUM(NSInteger, SPTPersistentCacheResponseCode) {
+    /**
+     * Indicates success of requested operation with data. The record field of SPTPersistentCacheResponse mustn't be nil
+     * if it was load operation otherwise it could be. The error would be nil.
+     */
+    SPTPersistentCacheResponseCodeOperationSucceeded,
+    /**
+     * Indicates that no file found for given key in cache or is expired. The record and error field of
+     * SPTPersistentCacheResponse is nil in this case.
+     */
+    SPTPersistentCacheResponseCodeNotFound,
+    /**
+     * Indicates error occured during requested operation. The record field of SPTPersistentCacheResponse would be nil.
+     * The error mustn't be nil and specify exact error.
+     */
+    SPTPersistentCacheResponseCodeOperationError
+};
 
 /**
  * @brief SPTPersistentCacheResponse
